+++ voicenavigator-20180612/src/Models/AuditTTSVoiceRequest.php
@@ -9,11 +9,36 @@ use AlibabaCloud\Tea\Model;
 class AuditTTSVoiceRequest extends Model
 {
     /**
+     * @var string
+     */
+    public $accessKey;
+
+    /**
+     * @var string
+     */
+    public $appKey;
+
+    /**
+     * @var string
+     */
+    public $engine;
+
+    /**
      * @example af81a389-91f0-4157-8d82-720edd02b66a
      *
      * @var string
      */
     public $instanceId;
+
+    /**
+     * @var string
+     */
+    public $pitchRate;
+
+    /**
+     * @var string
+     */
+    public $secretKey;
 
     /**
      * @example 100
@@ -41,7 +66,12 @@ class AuditTTSVoiceRequest extends Model
      */
     public $volume;
     protected $_name = [
+        'accessKey'  => 'AccessKey',
+        'appKey'     => 'AppKey',
+        'engine'     => 'Engine',
         'instanceId' => 'InstanceId',
+        'pitchRate'  => 'PitchRate',
+        'secretKey'  => 'SecretKey',
         'speechRate' => 'SpeechRate',
         'text'       => 'Text',
         'voice'      => 'Voice',
@@ -55,8 +85,23 @@ class AuditTTSVoiceRequest extends Model
     public function toMap()
     {
         $res = [];
+        if (null !== $this->accessKey) {
+            $res['AccessKey'] = $this->accessKey;
+        }
+        if (null !== $this->appKey) {
+            $res['AppKey'] = $this->appKey;
+        }
+        if (null !== $this->engine) {
+            $res['Engine'] = $this->engine;
+        }
         if (null !== $this->instanceId) {
             $res['InstanceId'] = $this->instanceId;
+        }
+        if (null !== $this->pitchRate) {
+            $res['PitchRate'] = $this->pitchRate;
+        }
+        if (null !== $this->secretKey) {
+            $res['SecretKey'] = $this->secretKey;
         }
         if (null !== $this->speechRate) {
             $res['SpeechRate'] = $this->speechRate;
@@ -82,8 +127,23 @@ class AuditTTSVoiceRequest extends Model
     public static function fromMap($map = [])
     {
         $model = new self();
+        if (isset($map['AccessKey'])) {
+            $model->accessKey = $map['AccessKey'];
+        }
+        if (isset($map['AppKey'])) {
+            $model->appKey = $map['AppKey'];
+        }
+        if (isset($map['Engine'])) {
+            $model->engine = $map['Engine'];
+        }
         if (isset($map['InstanceId'])) {
             $model->instanceId = $map['InstanceId'];
+        }
+        if (isset($map['PitchRate'])) {
+            $model->pitchRate = $map['PitchRate'];
+        }
+        if (isset($map['SecretKey'])) {
+            $model->secretKey = $map['SecretKey'];
         }
         if (isset($map['SpeechRate'])) {
             $model->speechRate = $map['SpeechRate'];